Create BankDriver.java
Create BankDriver class(this class has a main method in it):
  Create an array of thirty customers with data members customer name, number, balance and phone number.  

The bank class should have method called runBankTest() method to control the invocation of all the required methods.

First read all records from the input file customer.txt. Place the records in the customer array. Keep a count of the records placed in the array. Display appropriate messages such as: 
        a. NumberFormatException: “ Error in input. Line ignored.”
        b.  FileNotFoundException:     “The file xxx was not found.”
                     where xxx is customer.txt
c. IOException 
 Once all records are read and tabled, sort the array and display the records on 
the screen. Close the input file.
 Prompt the users to perform tasks
           1. Deposit sum to the account
           2. Withdraw sum from the account
           3. Create an account
           4. View all accounts
           5. Delete a customer’s account
           9. Quit Processing

Option 1. Find the index of the customer in the array. Call the customer’s  findIndex()  method . The method will return the index of the customer in the list, or return -1 if the customer is not in the array. Invoke the Customer’s deposit() method to deposit the amount in the savings account. Call addInterest() method to add 4.5 percent interest to the customer’s balance.

Option 2: Find the index of the customer in the array. Call the customer’s findIndex()  method . The method will return the index of the customer in the list, or return -1 if the customer is not in the array. Invoke the Customer’s withdraw() method  to withdraw  the amount from the savings account. Call addInterest() method to add 4.5 percent interest to the customer’s balance. 

Option 3: If the array is not full, invoke the Customer’s addNewCustomer method to add a new customer to the end of the array. 

Option 4: Sort all records in the array Call the customer’s nameSort() method.  Display the records on the screen. 

Option 5: Delete a customer’s record. Prompt the user to enter the name of the customer to delete. Validate that the customer is valid. Find the index of the customer in the array. Call the customer’s  findIndex() method . The method will return the index of the customer in the list, or return -1 if the customer is not in the array. If  a valid customer, call the Customer’s deleteCustomer( ) method to delete the customer.

Option 9: At end of processing, sort the customer’s array and display the records on the screen. Call the customer’s nameSort() method to sort the Customer Names and display